la-agent-py/pyproject.toml
[project]
name = "la-agent-py"
version = "0.1.0"
description = "LaAgent backend service"
readme = "readme.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "flask[async]==3.1.0",
  "flask-openapi3==4.1.0",
  "flask-openapi3-swagger==5.20.1",
  "flask-sqlalchemy==3.1.1",
  "flask-babel==4.0.0",
  "gevent==24.11.1",
  "langchain-ollama==0.2.3",
  "pydantic~=2.10.6",
  "requests~=2.32.3",
  "neo4j~=5.28.1",
  "flask-login~=0.6.3",
  "werkzeug~=3.1.3",
  "flask-cors~=5.0.1",
  "langchain-community~=0.3.20",
  "langchain-text-splitters~=0.3.7",
  "flask-sse~=1.0.0",
  "sentence-transformers~=4.0.1",
  "pyyaml~=6.0.2",
  "minio~=7.2.15",
  "redis~=5.2.1",
  "flask-redis~=0.4.0",
  "langgraph~=0.3.21",
  "langsmith~=0.3.19",
  "grandalf~=0.8",
  "langchain-mcp-adapters~=0.0.6",
  "flask-pydantic~=0.12.0",
  "mcp~=1.4.1",
  "loguru~=0.7.3",
  "bs4~=0.0.2",
  "jieba~=0.42.1",
  "celery~=5.5.0",
  "sqlalchemy-serializer~=1.4.22",
  "feign-python~=0.0.5",
  "toolbox~=1.11.0",
  "psycopg-binary~=3.2.6",
  "docx2txt~=0.9",
  "pymupdf~=1.25.5",
  "rapidocr-onnxruntime~=1.4.4",
  "referencing~=0.36.2",
  "unstructured~=0.17.2",
  "openpyxl~=3.1.5",
  "langchain-postgres~=0.0.14",
  "gunicorn~=23.0.0",
  "pyahocorasick~=2.1.0",
  "langchain-openai~=0.3.14",
  "langgraph-checkpoint~=2.0.24",
  "langgraph-checkpoint-postgres~=2.0.21",
  "pandas~=2.2.3",
  "langchain-experimental~=0.3.4",
  "polib~=1.2.0",
]

[tool.uv]
package = false

